Add tests for `From` coercion
